+++ packages/react/src/components/elements/customLayout.tsx
@@ -1,0 +1,12 @@
+import { cn } from "@/lib/utils";
+import React from "react";
+
+interface Props extends React.HTMLAttributes<HTMLDivElement> {}
+
+export const CustomLayout = ({ children, ...props }: Props) => {
+  return (
+    <div {...props} className={cn("w-full")}>
+      {children}
+    </div>
+  );
+};