+++ packages/react/src/components/elements/index.ts
@@ -12,3 +12,4 @@ export * from "./emailInput";
 export * from "./dateRangePicker";
 export * from "./numberInput";
 export * from "./phoneNumberInput";
+export * from "./customLayout";