+++ packages/react/src/components/elements/phoneNumberInput/phoneNumberInput.tsx
@@ -1,17 +1,37 @@
 import React from "react";
 import { Input as ShadInput } from "../../ui/input";
+import { cn } from "@/lib/utils";
+
+export interface PhoneNumberInputStyles {
+  root: React.CSSProperties;
+}
+
+export interface PhoneNumberInputClassNames {
+  root: string;
+}
 
 export interface PhoneNumberInputProps
   extends React.InputHTMLAttributes<HTMLInputElement> {
   isRequired?: boolean;
   label?: string;
+  styles?: Partial<PhoneNumberInputStyles>;
+  classNames?: Partial<PhoneNumberInputClassNames>;
 }
 
 export const PhoneNumberInput = ({
   isRequired,
+  styles,
+  classNames,
   ...props
 }: PhoneNumberInputProps) => {
   return (
-    <ShadInput required={isRequired} {...props} type="tel" inputMode="tel" />
+    <ShadInput
+      required={isRequired}
+      className={cn(classNames?.root)}
+      style={styles?.root}
+      {...props}
+      type="tel"
+      inputMode="tel"
+    />
   );
 };