+++ packages/react/src/components/elements/repeaterGroup/repeaterGroup.tsx
@@ -4,11 +4,23 @@ import { RepeaterGroupWrapper, RepeaterGroupWrapperProps } from "./wrapper";
 import { ArrayPath, FieldValues, UseFieldArrayReturn } from "react-hook-form";
 import { Trash } from "lucide-react";
 
+export interface RepeaterGroupStyles {
+  root: React.CSSProperties;
+  deleteButton: React.CSSProperties;
+}
+
+export interface RepeaterGroupClassNames {
+  root: string;
+  deleteButton: string;
+}
+
 interface RepeaterGroupProps<T extends FieldValues>
   extends DetailedHTMLProps<HTMLAttributes<HTMLDivElement>, HTMLDivElement> {
   onDelete?: () => unknown;
   readonly?: boolean;
   fields: UseFieldArrayReturn<T, ArrayPath<T>, "_key">["fields"];
+  styles?: Partial<RepeaterGroupStyles>;
+  classNames?: Partial<RepeaterGroupClassNames>;
 }
 
 export function RepeaterGroup<T extends FieldValues>({
@@ -17,6 +29,8 @@ export function RepeaterGroup<T extends FieldValues>({
   onDelete,
   readonly,
   fields,
+  styles,
+  classNames,
   ...props
 }: RepeaterGroupProps<T>) {
   return (
@@ -24,7 +38,9 @@ export function RepeaterGroup<T extends FieldValues>({
       className={cn(
         "w-full flex flex-col gap-2 p-2 border border-lfui-muted rounded-md relative",
         className,
+        classNames?.root,
       )}
+      style={styles?.root}
       {...props}
     >
       {children}
@@ -32,7 +48,11 @@ export function RepeaterGroup<T extends FieldValues>({
         <button
           type="button"
           onClick={onDelete}
-          className="px-2 py-1 text-xs rounded-sm bg-red-100 text-red-500 absolute top-1 right-1"
+          className={cn(
+            "px-2 py-1 text-xs rounded-sm bg-red-100 text-red-500 absolute top-1 right-1",
+            classNames?.deleteButton,
+          )}
+          style={styles?.deleteButton}
         >
           <Trash className="text-xs" size={14} />
         </button>