+++ packages/types/src/elements/input/index.ts
@@ -4,11 +4,17 @@ import {
   LezzformElementCategory,
 } from "../..";
 import { StringRule } from "../../rules";
+import { GeneralStyleMetadata } from "../../styles";
 
-export type InputType = GenericLezzformElement<GenericFieldAttributes> & {
+export type InputType = GenericLezzformElement<
+  GenericFieldAttributes,
+  InputStyleType
+> & {
   type: "Input";
   category: LezzformElementCategory.Field;
   rule: InputRuleType;
 };
+
+export type InputStyleType = GeneralStyleMetadata;
 
 export type InputRuleType = StringRule;