+++ packages/types/src/elements/textArea/index.ts
@@ -4,14 +4,18 @@ import {
   LezzformElementCategory,
 } from "../..";
 import { StringRule } from "../../rules";
+import { GeneralStyleMetadata } from "../../styles";
 
 export type TextAreaType = GenericLezzformElement<
-  GenericFieldAttributes & TextAreaAttributesType
+  GenericFieldAttributes & TextAreaAttributesType,
+  TextAreaStyleType
 > & {
   type: "TextArea";
   category: LezzformElementCategory.Field;
   rule: TextAreaRuleType;
 };
+
+export type TextAreaStyleType = GeneralStyleMetadata;
 
 export type TextAreaRuleType = StringRule;