+++ packages/types/src/elements/threeColumn/index.ts
@@ -1,6 +1,12 @@
 import { GenericLezzformElement, LezzformElementCategory } from "../..";
+import { GeneralStyleMetadata } from "../../styles";
 
-export type ThreeColumnType = GenericLezzformElement & {
+export type ThreeColumnType = GenericLezzformElement<
+  {},
+  ThreeColumnStyleType
+> & {
   type: "ThreeColumn";
   category: LezzformElementCategory.Layout;
 };
+
+export type ThreeColumnStyleType = GeneralStyleMetadata;